Remove compute_v
