+++ src/bots/pauser.py
@@ -9,7 +9,6 @@ from web3.types import BlockData
 import variables
 from blockchain.executor import Executor
 from blockchain.typings import Web3
-from cryptography.verify_signature import compute_vs
 from metrics.metrics import UNEXPECTED_EXCEPTIONS
 from metrics.transport_message_metrics import message_metrics_filter
 from transport.msg_providers.kafka import KafkaMessageProvider
@@ -121,7 +120,7 @@ class PauserBot:
             return False
 
         pause_tx = self.w3.lido.deposit_security_module.pause_deposits(
-            message['blockNumber'], module_id, (message['signature']['r'], compute_vs(message['signature']['v'], message['signature']['s']))
+            message['blockNumber'], module_id, (message['signature']['r'], message['signature']['_vs'])
         )
 
         if not self.w3.transaction.check(pause_tx):
@@ -138,7 +137,7 @@ class PauserBot:
             return False
 
         pause_tx = self.w3.lido.deposit_security_module.pause_deposits_v2(
-            message['blockNumber'], (message['signature']['r'], compute_vs(message['signature']['v'], message['signature']['s']))
+            message['blockNumber'], (message['signature']['r'], message['signature']['_vs'])
         )
 
         result = self.w3.transaction.send(pause_tx, False, 6)
